+++ CHANGES.rst
@@ -12,8 +12,8 @@ Changelog
   [mamico]
 - Subject templating
   [folix-01]
-- Nothing changed yet.
-
+- Do not set values in __init__ in *SubmitPost* because the user there is not already set and can lead to problems.
+  [cekk]
 
 3.2.3 (2025-03-07)
 ------------------